Change so that Block files are deleted immediately upon reading
They were already being deleted at the end of the run.
Now they are deleted as read, which leaves room for the ic files.
This cuts the disk footprint by 1/3.

This was done with a hardcoded option to the BlockArray, so one could
change this to a user-option in the future.  But there's no active use case
for that option.  It would only matter if we wanted to be able to restart
a failed job without regenerating the Block files, but that's only 50% of
the time on a not-that-expensive code.